+++ pyccel/codegen/printing/ccode.py
@@ -3,9 +3,11 @@
 # This file is part of Pyccel which is released under MIT License. See the LICENSE file or #
 # go to https://github.com/pyccel/pyccel/blob/devel/LICENSE for full license details.      #
 #------------------------------------------------------------------------------------------#
+import ast
 import functools
 from itertools import chain, product
 import re
+import sys
 from packaging.version import Version
 
 import numpy as np
@@ -669,9 +671,8 @@ class CCodePrinter(CodePrinter):
         expr : TypedAstNode
             The object representing the container being printed (e.g., PythonList, PythonSet).
 
-        assignment_var : Assign
-            The assignment node where the Python container (rhs) is being initialized
-            and saved into a variable (lhs).
+        assignment_var : Variable
+            The variable that the Python container is being assigned to.
 
         Returns
         -------
@@ -679,14 +680,14 @@ class CCodePrinter(CodePrinter):
             The generated C code for the container initialization.
         """
 
-        class_type = assignment_var.lhs.class_type
+        class_type = assignment_var.class_type
         dtype = self.get_c_type(class_type)
         if isinstance(expr, PythonDict):
             dict_item_strs = [(self._print(k), self._print(v)) for k,v in zip(expr.keys, expr.values)]
             keyraw = '{' + ', '.join(f'{{{k}, {v}}}' for k,v in dict_item_strs) + '}'
         else:
             keyraw = '{' + ', '.join(self._print(a) for a in expr.args) + '}'
-        container_name = self._print(assignment_var.lhs)
+        container_name = self._print(assignment_var)
         init = f'{container_name} = c_init({dtype}, {keyraw});\n'
         return init
 
@@ -2392,7 +2393,7 @@ class CCodePrinter(CodePrinter):
             return self.arrayFill(expr)
         lhs_code = self._print(lhs)
         if isinstance(rhs, (PythonList, PythonSet, PythonDict)):
-            return self.init_stc_container(rhs, expr)
+            return self.init_stc_container(rhs, expr.lhs)
         rhs_code = self._print(rhs)
         if isinstance(rhs, LiteralString):
             rhs_code = f'cstr_lit({rhs_code})'
@@ -2667,9 +2668,14 @@ class CCodePrinter(CodePrinter):
         return '/*' + comments + '*/\n'
 
     def _print_Assert(self, expr):
+        if isinstance(expr.test, LiteralTrue):
+            if sys.version_info < (3, 9):
+                return ''
+            else:
+                return '//' + ast.unparse(expr.python_ast) + '\n' #pylint: disable=no-member
         condition = self._print(expr.test)
         self.add_import(c_imports['assert'])
-        return "assert({0});\n".format(condition)
+        return f"assert({condition});\n"
 
     def _print_PyccelSymbol(self, expr):
         return expr
@@ -2829,6 +2835,12 @@ class CCodePrinter(CodePrinter):
         set_var = self._print(ObjectAddress(expr.set_variable))
         arg_val = self._print(expr.args[0])
         return f'{var_type}_erase({set_var}, {arg_val});\n'
+
+    def _print_PythonSet(self, expr):
+        tmp_var = self.scope.get_temporary_variable(expr.class_type, shape = expr.shape,
+                    memory_handling='heap')
+        self._additional_code += self.init_stc_container(expr, tmp_var)
+        return self._print(tmp_var)
 
     #================== Dict methods ==================